+++ client/Admin/index.js
@@ -1,7 +1,7 @@
 //Admin components
 export {default as AdminMain} from './main'
-export {default as AllOrders} from './AllOrders'
-export {default as AllOrdersModal} from './AllOrdersModal'
-export {default as AllOrdersTableRow} from './AllOrdersTableRow'
+export {default as Orders} from './Orders'
+export {default as OrdersModal} from './OrdersModal'
+export {default as OrdersTableRow} from './OrdersTableRow'
 export {default as AdminMenu} from './AdminMenu'
 export {default as AdminHeader} from './AdminHeader'
