+++ client/admin/Orders.js
@@ -2,9 +2,9 @@ import React from 'react'
 import { Table, Segment } from 'semantic-ui-react'
 import { connect } from 'react-redux'
 import { withRouter } from 'react-router-dom'
-import { AllOrdersModal } from './index'
+import { OrdersModal } from './index'
 
-const AllOrders = (props) => {
+const Orders = (props) => {
   return (
   <Segment attached="bottom">
     <Table selectable celled padded attached="bottom">
@@ -20,7 +20,7 @@ const AllOrders = (props) => {
       </Table.Header>
       <Table.Body>
         {props.users.length ? props.orders.map(order =>
-           <AllOrdersModal key={order.id} order={order} users={props.users} />
+           <OrdersModal key={order.id} order={order} users={props.users} />
         ) : null}
       </Table.Body>
     </Table>
@@ -29,4 +29,4 @@ const AllOrders = (props) => {
 
 const mapStateToProps = ({orders, users}) => ({orders, users})
 
-export default withRouter(connect(mapStateToProps)(AllOrders))
+export default withRouter(connect(mapStateToProps)(Orders))
